refactor(extractors): split synthetic_symbols into per-source package
Convert extractors/synthetic_symbols.py into a package with:
- __init__.py — façade preserving extract_synthetic_symbols import path,
  plus a multi-provider dispatcher (a language can register more than
  one provider — Lombok + records + MapStruct all run for Java).
- _helpers.py — shared build_synthetic_symbol and enclosing_type_name.
- csharp_mvvm.py — verbatim of the existing CommunityToolkit pass.

Fix a Java/Kotlin → C# regression introduced earlier: removed the
'parameter' entry from _PARAM_ORIGIN_BY_ANCESTOR — C# uses the same
node name for ctor/method params, so adding it overrode the walk to
constructor_declaration → ctor_param.
